kconfig: return imaginary newline if one is lacking at end of file
Before, on encountering EOF without EOL, we issued a warning and called
zconf_endfile() which freed current_buf and current_file structures.
But they may still be necessary to handle the last line. For example,
if the last line is source, then handling it involves a call to
zconf_nextfile() which expects valid current_buf. It seems handling
this properly will require a substantial redesign of the way things
currently work (it's not clear what other execution paths could be
tripped up by freed current_buf/file) so for now we first return an
imaginary EOL, which seems to do the trick.

Signed-off-by: Boris Kolpackov <boris@codesynthesis.com>