kconfig: turn lack of newline at end of file into error
Before it used to be a warning, however, on encountering EOF we call
zconf_endfile() which frees current_buf and current_file structures.
But they may still be necessary to handle the last line. For example,
if the last line is source, then handling it involves a call to
zconf_nextfile() which expects valid current_buf. It seems handling
this properly will require a substantial redesign of the way things
currently work (it's not clear what other execution paths could be
tripped up by freed current_buf/file) so for now we make this an
error, which is better than SIGSEGV.

Signed-off-by: Boris Kolpackov <boris@codesynthesis.com>